Rename new mahalanobis -> _mahalanobis
